refactor: convert myinto-types.ts to index.ts
