+++ src/types/index.ts
@@ -131,3 +131,54 @@ type VerifyIPersonCheck<
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   Missing extends never = IPersonCheck
 > = never
+
+export * from './address'
+export * from './aliasname'
+export * from './base'
+export * from './birthcountry'
+export * from './childrenbirthrecords'
+export * from './countryofmarriage'
+export * from './cpfbalances'
+export * from './cpfcontributions'
+export * from './cpfemployers'
+export * from './dialect'
+export * from './divorcedate'
+export * from './dob'
+export * from './drivinglicence'
+export * from './email'
+export * from './employment'
+export * from './employmentsector'
+export * from './gstvoucher'
+export * from './hanyupinyinaliasname'
+export * from './hanyupinyinname'
+export * from './hdbownership'
+export * from './hdbtype'
+export * from './householdincome'
+export * from './housingtype'
+export * from './marital'
+export * from './marriagecertno'
+export * from './marriagedate'
+export * from './marriedname'
+export * from './merdekagen'
+export * from './mobileno'
+export * from './name'
+export * from './nationality'
+export * from './noa-basic'
+export * from './noa'
+export * from './noahistory-basic'
+export * from './noahistory'
+export * from './occupation'
+export * from './ownerprivate'
+export * from './passexpirydate'
+export * from './passportexpirydate'
+export * from './passportnumber'
+export * from './passstatus'
+export * from './passtype'
+export * from './race'
+export * from './residentialstatus'
+export * from './secondaryrace'
+export * from './sex'
+export * from './silversupport'
+export * from './sponsoredchildrenrecords'
+export * from './uinfin'
+export * from './vehicles'
